Initial commit!
Idea for the project is to create a script that will detect if a file was added
parse the file and update the corresponding values in the server config.
I call it a config bomb because thats what came to mind when thinking about the action of
uploading it and having it take effect.